mm: memcg: update the correct soft limit tree during migration
end_migration() passes the old page instead of the new page to commit the
charge.  This page descriptor is not used for committing itself, though,
since we also pass the (correct) page_cgroup descriptor.  But it's used to
find the soft limit tree through the page's zone, so the soft limit tree
of the old page's zone is updated instead of that of the new page's, which
might get slightly out of date until the next charge reaches the ratelimit
point.

This glitch has been present since '5564e88 memcg: condense
page_cgroup-to-page lookup points'.

This fixes a bug that I introduced in 2.6.38.  It's benign enough (to
my knowledge) that we probably don't want this for stable.

Reported-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Cc: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>
Acked-by: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>