Fix up for: mm: remove lru type checks from __isolate_lru_page()
I screwed when applying Konstantin's patch from mmotm without noticing
and missed this part of the patch because of some whitespace noise.
I would have rebased the patch (to hide this) but I already pushed the
tree so let's do it as a separate commit. I will fix this up during the
next rebase round when creating since-3.4 branch.

Sorry about this.