mm: sparse: fix usemap allocation above node descriptor section
After f5bf18f ("bootmem/sparsemem: remove limit constraint in
alloc_bootmem_section"), usemap allocations may easily be placed outside
the optimal section that holds the node descriptor, even if there is space
available in that section.  This results in unnecessary hotplug
dependencies that need to have the node unplugged before the section
holding the usemap.

The reason is that the bootmem allocator doesn't guarantee a linear search
starting from the passed allocation goal but may start out at a much
higher address absent an upper limit.

Fix this by trying the allocation with the limit at the section end, then
retry without if that fails.  This keeps the fix from f5bf18f of not
panicking if the allocation does not fit in the section, but still makes
sure to try to stay within the section at first.

Signed-off-by: Yinghai Lu <yinghai@kernel.org>
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Cc: <stable@vger.kernel.org>	[3.3.x, 3.4.x]
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>