parisc: redefine ATOMIC_INIT and ATOMIC64_INIT like other architectures
The following build error occured during a parisc build with
swap-over-NFS patches applied.

net/core/sock.c:274:36: error: initializer element is not constant
net/core/sock.c:274:36: error: (near initialization for 'memalloc_socks')
net/core/sock.c:274:36: error: initializer element is not constant

It's not obvious but this is due to how ATOMIC_INIT is defined on parisc.
It should affect any user of STATIC_KEY_INIT_FALSE on that platform.

This patch makes the definition of ATOMIC_INIT on parisc to look like
other arches definition.

Apparently due to gcc creating a temporary.  John said "Parisc is a callee
copies architecture.  Thus, a structure passed as an argument may be
copied to a temporary.  This has been a source of numerous bugs as the
majority of architectures do it the other way round."

Reported-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Mel Gorman <mgorman@suse.de>
Cc: James Bottomley <James.Bottomley@hansenpartnership.com>
Cc: John David Anglin <dave.anglin@bell.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>